Revert "Changed the Position.java datatype"
This reverts commit 0c4007bf49db2e8a04e1ddb6d6c1a10a0981d02c.

+++ Position.java
@@ -5,8 +5,6 @@
  *
  */
 
-import java.math.BigDecimal;
-
 /**
  * Position is a class which defines a datatype Position which contains an x and y coord.
  *
@@ -17,15 +15,15 @@ public class Position {
 	 * the x coord for a position.
 	 * the y coord for a position.
 	 */
-	private BigDecimal positionX;
-	private BigDecimal positionY;
+	private double positionX;
+	private double positionY;
 	
 	/**
 	 * This defines a position of x and y.
 	 * @param x
 	 * @param y
 	 */
-	public Position (BigDecimal x, BigDecimal y) {
+	public Position (double x, double y) {
 		this.positionX = x;
 		this.positionY = y;
 	}
@@ -34,15 +32,15 @@ public class Position {
 	 * Returns the x and y coord as an array of doubleegers.
 	 *	@return 1D array consisting of x and y.
 	 */
-	public BigDecimal[] getPosition () {
-		BigDecimal retrunArray[] = {positionX, positionY};
+	public double[] getPosition () {
+		double retrunArray[] = {positionX, positionY};
 		return retrunArray;
 	}
 	
 	/**
 	 * Overwrites the current x and y of a position,
 	 */
-	public void setPosition (BigDecimal x, BigDecimal y) {
+	public void setPosition (double x, double y) {
 		this.positionX = x;
 		this.positionY = y;
 	}
